requirements.txt
langchain
langchain_openai
langchain_community
python-dotenv
pyodbc
streamlit
pandas
openpyxl
